requirements.txt
streamlit>=1.38.0
openai>=1.40.0
plotly>=5.24.0
psycopg2-binary>=2.9.9
pandas>=2.2.0
python-dotenv>=1.0.0
